feat: add documentation for Messages
We have introduced a new feature called notification feed, it requires
the employers to create messages they want to send to the their
employees.
We have thus introduced an API enpoint to support this and also added a
support to view it in the gigapay app.

The employer has full ownership over the message and change do what ever
they want at any time.